FastLED_Attempt1
Program creates 5 regions on the LED strip, 4 visually show how much power and current are running through sensors connected to a plug outlet. These 4 simulate 4 outlets in a room for example. The fifth region is a summation of these 4 other plug ins, to show how much power/current in total is going to or coming from this room.

A button is connected to turn the system off or on, and there are magnetic sensors checking whether a plug is in the outlet or not.

Problem:
1. Need to figure out cleaner way of creating the sine wave on the LED strip with less code.
2. Need to create a message function that can show a green light starting at the first LED of each of the 4 region's, which moves to the end of the LED strip, once the green light reaches the end for 1 of the respected regions, the sinewave will begin at the region which reached the end.